event_finder/requirements_for_lambda.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.12.3
boto3==1.37.29
botocore==1.37.29
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.3.2
propcache==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
